fix added for too short videos

+++ NCG_Full/NCG_VideoEditor.cs
@@ -73,7 +73,7 @@ namespace NCG_Full
                     break;
                 }
 
-                if (songDuration < 120) //delete file is duration is less than 2min
+                if (songDuration > 0 && songDuration < 120) //delete file is duration is less than 2min
                 {
                     Console.WriteLine("Song too short, deleting file...");
                     File.Delete(songPath);
